+++ src/defaults.vala
@@ -12,4 +12,7 @@ namespace Peek {
   const int DEFAULT_FRAMERATE = 15;
   const int DEFAULT_DOWNSAMPLE = 1;
 
+  const string OUTPUT_FORMAT_GIF = "gif";
+  const string OUTPUT_FORMAT_WEBM = "webm";
+
 }