+++ src/recording/ffmpeg-screen-recorder.vala
@@ -19,29 +19,58 @@ namespace Peek.Recording {
         // Cancel running recording
         cancel ();
 
-        temp_file = Utils.create_temp_file ("avi");
         string display = Environment.get_variable ("DISPLAY");
         if (display == null) {
           display = ":0";
         }
 
-        string[] args = {
-          "ffmpeg", "-y",
-          // "-loglevel", "debug",
-          "-f", "x11grab",
-          "-show_region", "0",
-          "-framerate", framerate.to_string (),
-          "-video_size", area.width.to_string () + "x" + area.height.to_string (),
-          // "-use_shm", "0",
-          // "-draw_mouse", "0",
-          "-i", display + "+" + area.left.to_string () + "," + area.top.to_string (),
-          "-codec:v", "libx264rgb",
-          "-preset:v", "ultrafast",
-          "-filter:v", "crop=iw-mod(iw\\,2):ih-mod(ih\\,2), scale=iw/" + downsample.to_string () + ":-1",
-          temp_file
-        };
+        var args = new Array<string> ();
+        args.append_val ("ffmpeg");
+        args.append_val ("-y");
+        // args.append_val ("-loglevel");
+        // args.append_val ("debug");
+        args.append_val ("-f");
+        args.append_val ("x11grab");
+        args.append_val ("-show_region");
+        args.append_val ("0");
+        args.append_val ("-framerate");
+        args.append_val (framerate.to_string ());
+        args.append_val ("-video_size");
+        args.append_val (area.width.to_string () + "x" + area.height.to_string ());
+        // args.append_val ("-draw_mouse");
+        // args.append_val ("0");
+        args.append_val ("-i");
+        args.append_val (display + "+" + area.left.to_string () + "," + area.top.to_string ());
+        args.append_val ("-codec:v");
 
-        return spawn_record_command (args);
+        string extension;
+
+        if (output_format == OUTPUT_FORMAT_WEBM) {
+          extension = "webm";
+          // args.append_val ("libvpx-vp9");
+          args.append_val ("libvpx");
+          args.append_val ("-qmin");
+          args.append_val ("10");
+          args.append_val ("-qmax");
+          args.append_val ("50");
+          args.append_val ("-crf");
+          args.append_val ("13");
+          args.append_val ("-b:v");
+          args.append_val ("1M");
+        } else {
+          extension = "avi";
+          args.append_val ("libx264rgb");
+          args.append_val ("-preset:v");
+          args.append_val ("ultrafast");
+        }
+
+        args.append_val ("-filter:v");
+        args.append_val ("crop=iw-mod(iw\\,2):ih-mod(ih\\,2), scale=iw/" + downsample.to_string () + ":-1");
+
+        temp_file = Utils.create_temp_file (extension);
+        args.append_val (temp_file);
+
+        return spawn_record_command (args.data);
       } catch (FileError e) {
         stderr.printf ("Error: %s\n", e.message);
         return false;
@@ -52,7 +81,7 @@ namespace Peek.Recording {
       return Utils.check_for_executable ("ffmpeg");
     }
 
-    protected override void stop_command () {
+    protected override void stop_recording () {
       try {
         char[] command = { 'q' };
         size_t bytes_written;