+++ src/recording/screen-recorder.vala
@@ -12,6 +12,8 @@ namespace Peek.Recording {
   public interface ScreenRecorder : Object {
     public abstract bool is_recording { get; protected set; }
 
+    public abstract string output_format { get; set; }
+
     public abstract int framerate { get; set; }
 
     public abstract int downsample { get; set; }