+++ CHANGES.md
@@ -1,6 +1,8 @@
 # Version 0.10.0 - 2017-??-??
  * recording: Support Gnome Shell screencast DBUS service. Allows recording
-   under Gnome Wayland.
+   under Gnome Wayland (#33).
+ * recording: Support WebM as output format (#73)
+ * fix: Fixd a possible race condition that could lead to empty or broken files (#1)
 
 # Version 0.9.0 - 2017-02-21
  * i18n: Fixed Czech, Croatian, Korean, Dutch and Chinese (Simplified) not getting installed